Backend/Python/requirements.txt
boto3==1.10.44
botocore==1.13.44
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.4
psycopg2==2.8.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
